Update ⁨اوامر الكروب.py⁩

+++ JoKeRUB/plugins/اوامر الكروب.py
@@ -829,7 +829,10 @@ delete_enabled = False
 @l313l.on(events.NewMessage)
 async def Hussein(event):
     global delete_enabled
-    if event.is_private and delete_enabled:
+    sender_id = event.sender_id
+    current_user_id = await client.get_me()
+    
+    if event.is_private and delete_enabled and sender_id == current_user_id.id:
         await event.delete()
 @l313l.ar_cmd(pattern="الخاص تعطيل")
 async def joker5a9(event: Message):
